feat/germinare/useRouter
